feat: history store
