+++ src/components/Header.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import { storeToRefs } from 'pinia'
-import useAuthStore from '@/stores/auth'
-import { useRoute, useRouter } from 'vue-router'
 import { ref, onMounted } from 'vue'
+import { storeToRefs } from 'pinia'
+import { useRoute, useRouter } from 'vue-router'
+import useAuthStore from '@/stores/auth'
 
 const route = useRoute()
 const router = useRouter()
