+++ src/components/ToolBar/UndoRedoButton.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { reactive } from 'vue'
-import useStoreStage from '@/stores/konva/stage'
+import useStoreHistory from '@/stores/konva/history'
 
-const { handleUndo, handleRedo } = useStoreStage()
+const { handleUndo, handleRedo } = useStoreHistory()
 
 const undoRedoArray = reactive([
   {
