+++ src/stores/konva/history.ts
@@ -12,57 +12,14 @@ interface History {
   images: KonvaImage[]
 }
 
-const useStoreStage = defineStore({
-  id: 'stage',
+const useStoreHistory = defineStore({
+  id: 'history',
   state: () => ({
-    configKonva: {
-      size: {
-        width: window.innerWidth,
-        height: window.innerWidth,
-      },
-      scale: {
-        x: 1,
-        y: 1,
-      },
-    },
     historyStep: 0,
     canvasHistory: [{ lines: [], texts: [], images: [] }] as History[],
   }),
 
   actions: {
-    fitStageIntoParentContainer(container: HTMLDivElement) {
-      // Fixed stage size
-      const SCENE_BASE_WIDTH = 896
-      const SCENE_BASE_HEIGHT = 504
-
-      // Max upscale
-      const SCENE_MAX_WIDTH = 1280
-      const SCENE_MAX_HEIGHT = 720
-
-      if (container === null) return
-
-      const stageWidth =
-        container.offsetWidth % 2 !== 0
-          ? container.offsetWidth - 1
-          : container.offsetWidth
-
-      this.configKonva.size = {
-        width: stageWidth,
-        height: (stageWidth * 9) / 16, // aspect-ratio
-      }
-
-      const scaleX =
-        Math.min(this.configKonva.size.width, SCENE_MAX_WIDTH) /
-        SCENE_BASE_WIDTH
-
-      const scaleY =
-        Math.min(this.configKonva.size.height, SCENE_MAX_HEIGHT) /
-        SCENE_BASE_HEIGHT
-
-      const minRatio = Math.min(scaleX, scaleY)
-      this.configKonva.scale = { x: minRatio, y: minRatio }
-    },
-
     handleEventEndSaveHistory() {
       const { lines } = storeToRefs(useStoreLine())
       const { texts } = storeToRefs(useStoreText())
@@ -128,4 +85,4 @@ const useStoreStage = defineStore({
   },
 })
 
-export default useStoreStage
+export default useStoreHistory
