+++ src/stores/konva/text.ts
@@ -3,7 +3,7 @@ import { defineStore, storeToRefs } from 'pinia'
 import { nanoid } from 'nanoid'
 import Konva from 'konva'
 import useStoreMode from '@/stores/mode'
-import useStoreStage from '@/stores/konva/stage'
+import useStoreHistory from '@/stores/konva/history'
 import useStoreTransformer from '@/stores/konva/transformer'
 
 type FontStyle = 'normal' | 'bold' | 'italic' | 'italic bold'
@@ -92,7 +92,7 @@ const useStoreText = defineStore({
           name: 'text',
         },
       ]
-      useStoreStage().handleEventEndSaveHistory()
+      useStoreHistory().handleEventEndSaveHistory()
     },
 
     setTextOptionValue(option: string, value: string | TextAlign) {
@@ -115,7 +115,7 @@ const useStoreText = defineStore({
           default:
             break
         }
-        useStoreStage().handleEventEndSaveHistory()
+        useStoreHistory().handleEventEndSaveHistory()
       }
     },
 
@@ -249,7 +249,7 @@ const useStoreText = defineStore({
         this.removeTextarea(textNode, transformerNode, textarea)
         // 編集終了
         this.isEditing = false
-        useStoreStage().handleEventEndSaveHistory()
+        useStoreHistory().handleEventEndSaveHistory()
       })
     },
 
@@ -272,7 +272,7 @@ const useStoreText = defineStore({
         text.x = shape.x()
         text.y = shape.y()
       }
-      useStoreStage().handleEventEndSaveHistory()
+      useStoreHistory().handleEventEndSaveHistory()
     },
   },
 })
