+++ src/views/SignUpPage.vue
@@ -2,7 +2,7 @@
 import { reactive } from 'vue'
 import router from '@/router/index'
 import useAuthStore from '@/stores/auth'
-// import { InputTextType } from '@/types/index'
+
 interface InputTextType {
   icon: string
   inputType: string
